requirements.txt

transformers==4.44.2
torch>=2.2.0
datasets>=2.19.0
accelerate>=0.33.0
scikit-learn>=1.4.0
numpy
pandas
captum>=0.7.0
shap>=0.45.0
beautifulsoup4
tldextract
idna
mailparser
python-multipart
python-dotenv
fastapi
uvicorn[standard]
pydantic
streamlit>=1.33.0
pyyaml
httpx
pytest
pytest-asyncio
ruff
black
